fix some bugs (#1443)
* fix: error in reset_all_users_data_usage

* fix: make on_hold fields None when user is activate

* fix: Add last_status_change to disable/activate_all_active_users

+++ app/db/crud.py
@@ -662,8 +662,9 @@ def reset_all_users_data_usage(db: Session, admin: Optional[Admin] = None):
             dbuser.status = UserStatus.active
         dbuser.usage_logs.clear()
         dbuser.node_usages.clear()
-        db.delete(dbuser.next_plan)
-        dbuser.next_plan = None
+        if dbuser.next_plan:
+            db.delete(dbuser.next_plan)
+            dbuser.next_plan = None
         db.add(dbuser)
 
     db.commit()
@@ -681,7 +682,7 @@ def disable_all_active_users(db: Session, admin: Optional[Admin] = None):
     if admin:
         query = query.filter(User.admin == admin)
 
-    query.update({User.status: UserStatus.disabled}, synchronize_session=False)
+    query.update({User.status: UserStatus.disabled, User.last_status_change: datetime.utcnow()}, synchronize_session=False)
 
     db.commit()
 
@@ -704,8 +705,10 @@ def activate_all_disabled_users(db: Session, admin: Optional[Admin] = None):
         query_for_active_users = query_for_active_users.filter(User.admin == admin)
         query_for_on_hold_users = query_for_on_hold_users.filter(User.admin == admin)
 
-    query_for_on_hold_users.update({User.status: UserStatus.on_hold}, synchronize_session=False)
-    query_for_active_users.update({User.status: UserStatus.active}, synchronize_session=False)
+    query_for_on_hold_users.update(
+        {User.status: UserStatus.on_hold, User.last_status_change: datetime.utcnow()}, synchronize_session=False)
+    query_for_active_users.update(
+        {User.status: UserStatus.active, User.last_status_change: datetime.utcnow()}, synchronize_session=False)
 
     db.commit()
 
@@ -811,6 +814,9 @@ def update_user_status(db: Session, dbuser: User, status: UserStatus) -> User:
     Returns:
         User: The updated user object.
     """
+    if dbuser.status == UserStatus.on_hold and status == UserStatus.active:
+        dbuser.on_hold_expire_duration = None
+        dbuser.on_hold_timeout = None
     dbuser.status = status
     dbuser.last_status_change = datetime.utcnow()
     db.commit()
